refactor: 유효성 검사 수정
